requirements.txt
packaging
ninja
wheel
setuptools
setuptools-scm
adam-atan2-pytorch
einops
tqdm
coolname
pydantic
argdantic
wandb
omegaconf
hydra-core
huggingface_hub
